:recycle: moving to utils folder
